Fix State Restoration on rotation bug
State restoration failed because the fragment was created twice thus
the viewmodel which survived rotation was replaced with a new viewmodel.
Interesting to note, is that super.onCreate()'s call on the Activity
restores fragments on configuration change, thus its not necessary to call
replaceFragment extension unless we're sure the activity is not being created
as a result of an orientation change.

Additionally, remove unnecessary hacks using SingleLiveEvent to fix the bug.
Upgrade lifecycle version.