Migrate to Coroutines
Since this app does not make use of a reactive model, i've opted to use
coroutines instead of RxJava as a solution for implementing asynchronous
programming.

Additionally, i've added factory methods to UiModel to simplify the creation
of UiModel's in specific states. This improves code readability at the
call-site.

Improve naming: changed ModelState to UiState for improved clarity.