Simplify repository implementation (#1)
For this use-case, the remote data source class provided no actual benefit,
it just added more complexity thus making development a tad bit slower.
For a cleaner mental model and faster development, while still maintaining
the benefits of the repository pattern, i've opted to remove it and create
an interface to represent the repository. Using dagger a concrete implementation
of this interface will be provided to the object graph.